// libs/glfw/src/enums.ts
export enum GLFW_ErrorCodes {
  /** No error has occurred. */
  NO_ERROR = 0,

  /** GLFW has not been initialized. */
  NOT_INITIALIZED = 0x00010001,

  /** No context is current for this thread. */
  NO_CURRENT_CONTEXT = 0x00010002,

  /** One of the arguments to the function was an invalid enum value. */
  INVALID_ENUM = 0x00010003,

  /** One of the arguments to the function was an invalid value. */
  INVALID_VALUE = 0x00010004,

  /** A memory allocation failed. */
  OUT_OF_MEMORY = 0x00010005,

  /** GLFW could not find support for the requested API on the system. */
  API_UNAVAILABLE = 0x00010006,

  /** The requested OpenGL or OpenGL ES version is not available. */
  VERSION_UNAVAILABLE = 0x00010007,

  /** A platform-specific error occurred that does not match any of the more specific categories. */
  PLATFORM_ERROR = 0x00010008,

  /** The requested format is not supported or available. */
  FORMAT_UNAVAILABLE = 0x00010009,

  /** The specified window does not have an OpenGL or OpenGL ES context. */
  NO_WINDOW_CONTEXT = 0x0001000a,

  /** The specified cursor shape is not available. */
  CURSOR_UNAVAILABLE = 0x0001000b,

  /** The requested feature is not provided by the platform. */
  FEATURE_UNAVAILABLE = 0x0001000c,

  /** The requested feature is not implemented for the platform. */
  FEATURE_UNIMPLEMENTED = 0x0001000d,

  /** Platform unavailable or no matching platform was found. */
  PLATFORM_UNAVAILABLE = 0x0001000e,
}